Added a result of "sample number" to processor
Required clarification of verifyAndAddMeasurementResult for int - overload now specifies int16_t vs uint32_t

+++ src/sensors/MaxBotixSonar.cpp
@@ -107,7 +107,7 @@ bool MaxBotixSonar::addSingleMeasurementResult(void)
     // Initialize values
     bool success = false;
     int rangeAttempts = 0;
-    int result = -9999;
+    int16_t result = -9999;
 
     // Clear anything out of the stream buffer
     int junkChars = _stream->available();
